sw-6589: Moves around more commands to custom chains
