+++ cypress/e2e/supervision/dashboard/reassign-client.cy.js
@@ -30,8 +30,8 @@ before(function setupAllocatedClient() {
   cy.loginAs("Allocations User");
   cy.createClient()
     .withOrder();
-  cy.get("@orderId").then((orderId) => cy.setSupervisionLevel(orderId));
-  cy.get("@orderId").then((orderId) => cy.changeOrderStatus(orderId));
+  cy.get("@orderId").then((orderId) => cy.withSupervisionLevel(orderId));
+  cy.get("@orderId").then((orderId) => cy.withOrderStatus(orderId));
   allocateTheClientToLayTeam();
 });
