+++ cypress/e2e/supervision/orders/create-order.cy.js
@@ -1,3 +1,75 @@
+const createOrder = (orderType, orderSubType, orderDate, optional) => {
+  let orderDay = orderDate.split("/")[0];
+  let orderMonth = orderDate.split("/")[1];
+  let orderYear = orderDate.split("/")[2];
+  cy.get("@client").then(({id, caseRecNumber}) => {
+    cy.visit(`/supervision/#/clients/${id}`);
+    cy.contains("Create order").click();
+    cy.get("#orderType")
+      .closest(".fieldset")
+      .contains(orderType)
+      .click();
+    cy.contains("Order subtype")
+      .closest(".fieldset")
+      .find("Select")
+      .select(orderSubType);
+    cy.get('input[name="courtReference"]').clear();
+    cy.get('input[name="courtReference"]').type(caseRecNumber);
+    cy.get("#fIELDLABELSORDERDATE_day").clear();
+    cy.get("#fIELDLABELSORDERDATE_day").type(orderDay);
+    cy.get("#fIELDLABELSORDERDATE_month").clear();
+    cy.get("#fIELDLABELSORDERDATE_month").type(orderMonth);
+    cy.get("#fIELDLABELSORDERDATE_year").clear();
+    cy.get("#fIELDLABELSORDERDATE_year").type(orderYear);
+    cy.get("#orderReceivedDate_day").clear();
+    cy.get("#orderReceivedDate_day").type(orderDay);
+    cy.get("#orderReceivedDate_month").clear();
+    cy.get("#orderReceivedDate_month").type(orderMonth);
+    cy.get("#orderReceivedDate_year").clear();
+    cy.get("#orderReceivedDate_year").type(orderYear);
+    if (optional) {
+      cy.get("#orderIssueDate_day").clear();
+      cy.get("#orderIssueDate_day").type(orderDay);
+      cy.get("#orderIssueDate_month").clear();
+      cy.get("#orderIssueDate_month").type(orderMonth);
+      cy.get("#orderIssueDate_year").clear();
+      cy.get("#orderIssueDate_year").type(orderYear);
+      cy.get("#orderExpiryDate_day").clear();
+      cy.get("#orderExpiryDate_day").type(orderDay);
+      cy.get("#orderExpiryDate_month").clear();
+      cy.get("#orderExpiryDate_month").type(orderMonth);
+      cy.get("#orderExpiryDate_year").clear();
+      cy.get("#orderExpiryDate_year").type(orderYear);
+      cy.get("#clauseExpiryDate_day").clear();
+      cy.get("#clauseExpiryDate_day").type(orderDay);
+      cy.get("#clauseExpiryDate_month").clear();
+      cy.get("#clauseExpiryDate_month").type(orderMonth);
+      cy.get("#clauseExpiryDate_year").clear();
+      cy.get("#clauseExpiryDate_year").type(orderYear);
+      cy.get('input[name="orderTitle"]').clear();
+      cy.get('input[name="orderTitle"]').type("Test Order Title");
+      cy.contains("How have the deputy/deputies been appointed?")
+        .closest(".fieldset")
+        .contains("Sole")
+        .click();
+      cy.window()
+        .its("tinyMCE")
+        .its("activeEditor")
+        .its("initialized", {timeout: 2000});
+      cy.window().then((win) => {
+        const pastedata =
+          '<p class="MsoNormal" style="margin: 0cm 0cm 11.25pt; font-size: 12pt; font-family: Calibri, sans-serif; text-align: justify; background: white;"><span style="font-size: 10.5pt; font-family: &quot;Open Sans&quot;, sans-serif;">Test note</span></p>';
+        let editor = win.tinymce.activeEditor;
+        editor.dom.createRng();
+        editor.execCommand("mceInsertClipboardContent", false, {
+          content: pastedata,
+        });
+      });
+    }
+    cy.contains("Save & exit").click();
+  });
+}
+
 beforeEach(() => {
   cy.loginAs("Allocations User");
   cy.createClient();
@@ -13,7 +85,7 @@ describe(
         url: "/supervision-api/v1/bond-providers",
       }).as("bondProviderCall");
       cy.loginAs("System Admin");
-      cy.createOrder("Property & affairs", "New deputy", "01/01/2022", false);
+      createOrder("Property & affairs", "New deputy", "01/01/2022", false);
       cy.wait("@bondProviderCall").then(() => {
         cy.contains("Cancel").click();
       });
@@ -37,7 +109,7 @@ describe(
     });
 
     it("creates a supervised h&w order in supervision with mandatory fields as an allocations user", () => {
-      cy.createOrder("Health & welfare", "New deputy", "01/01/2022", false);
+      createOrder("Health & welfare", "New deputy", "01/01/2022", false);
       cy.get(".TABS_ORDERS").click();
       cy.get("#order-table")
         .find("tr")
@@ -63,7 +135,7 @@ describe(
         url: "/supervision-api/v1/bond-providers",
       }).as("bondProviderCall");
       cy.loginAs("System Admin");
-      cy.createOrder("Property & affairs", "New deputy", "01/01/2022", true);
+      createOrder("Property & affairs", "New deputy", "01/01/2022", true);
       cy.wait("@bondProviderCall").then(() => {
         cy.contains("Cancel").click();
       });
@@ -97,7 +169,7 @@ describe(
     });
 
     it("creates a non-supervised hw order in supervision with mandatory fields", () => {
-      cy.createOrder("Health & welfare", "Guardianship", "01/01/2022", true);
+      createOrder("Health & welfare", "Guardianship", "01/01/2022", true);
       cy.get(".TABS_ORDERS").click();
       cy.get("#order-table")
         .find("tr")
