+++ cypress/parallel-weights.json
@@ -1,1 +1,1 @@
-{"cypress/e2e/common/login.cy.js":{"time":6049,"weight":4},"cypress/e2e/lpa/cases/administration.cy.js":{"time":13104,"weight":9},"cypress/e2e/lpa/cases/allocate-case.cy.js":{"time":10170,"weight":7},"cypress/e2e/lpa/cases/change-status.cy.js":{"time":12270,"weight":8},"cypress/e2e/lpa/cases/complaints.cy.js":{"time":17807,"weight":12},"cypress/e2e/lpa/cases/create-event.cy.js":{"time":9886,"weight":6},"cypress/e2e/lpa/cases/create-relationship.cy.js":{"time":12189,"weight":8},"cypress/e2e/lpa/cases/documents.cy.js":{"time":10959,"weight":7},"cypress/e2e/lpa/cases/epa.cy.js":{"time":17322,"weight":12},"cypress/e2e/lpa/cases/investigations.cy.js":{"time":38520,"weight":26},"cypress/e2e/lpa/cases/lpa.cy.js":{"time":21028,"weight":14},"cypress/e2e/lpa/cases/search.cy.js":{"time":8443,"weight":5},"cypress/e2e/lpa/cases/tasks.cy.js":{"time":33591,"weight":23},"cypress/e2e/lpa/cases/timeline.cy.js":{"time":6717,"weight":4},"cypress/e2e/lpa/cases/warnings.cy.js":{"time":17094,"weight":11},"cypress/e2e/lpa/donors/donor.cy.js":{"time":18943,"weight":13},"cypress/e2e/lpa/donors/link.cy.js":{"time":12136,"weight":8},"cypress/e2e/supervision-deputy-hub/navigation.cy.js":{"time":8170,"weight":5},"cypress/e2e/supervision-firm-deputy-hub/navigation.cy.js":{"time":7826,"weight":5},"cypress/e2e/supervision/clients/add-client-search.cy.js":{"time":4415,"weight":3},"cypress/e2e/supervision/clients/add-event.cy.js":{"time":4806,"weight":3},"cypress/e2e/supervision/clients/add-new-client.cy.js":{"time":7234,"weight":5},"cypress/e2e/supervision/clients/allocate-client.cy.js":{"time":8287,"weight":5},"cypress/e2e/supervision/clients/documents/open-document.cy.js":{"time":23968,"weight":16},"cypress/e2e/supervision/clients/edit-client.cy.js":{"time":10895,"weight":7},"cypress/e2e/supervision/clients/edit-finance-details.cy.js":{"time":13776,"weight":9},"cypress/e2e/supervision/clients/record-client-death-confirmation.cy.js":{"time":16149,"weight":11},"cypress/e2e/supervision/clients/risk-score.cy.js":{"time":8317,"weight":5},"cypress/e2e/supervision/clients/visits/add-visit.cy.js":{"time":9119,"weight":6},"cypress/e2e/supervision/clients/visits/edit-visit.cy.js":{"time":9113,"weight":6},"cypress/e2e/supervision/clients/visits/extend-visit.cy.js":{"time":11395,"weight":7},"cypress/e2e/supervision/dashboard/client-dashboard.cy.js":{"time":7236,"weight":5},"cypress/e2e/supervision/dashboard/dashboard.cy.js":{"time":4073,"weight":2},"cypress/e2e/supervision/dashboard/reassign-client.cy.js":{"time":10439,"weight":7},"cypress/e2e/supervision/deputies/add-deputy.cy.js":{"time":71197,"weight":49},"cypress/e2e/supervision/documents/download-multiple.cy.js":{"time":10455,"weight":7},"cypress/e2e/supervision/help-and-guidance/help-and-guidance.cy.js":{"time":4494,"weight":3},"cypress/e2e/supervision/letters/create-letter.cy.js":{"time":23729,"weight":16},"cypress/e2e/supervision/navigation.cy.js":{"time":4655,"weight":3},"cypress/e2e/supervision/orders/create-order.cy.js":{"time":63082,"weight":43},"cypress/e2e/supervision/search/full-search.cy.js":{"time":6700,"weight":4},"cypress/e2e/supervision/search/search-bar.cy.js":{"time":15169,"weight":10},"cypress/e2e/user-management/my-details.cy.js":{"time":2726,"weight":1},"cypress/e2e/user-management/navigation.cy.js":{"time":6212,"weight":4},"cypress/e2e/workflow/navigation.cy.js":{"time":8918,"weight":6}}
+{"cypress/e2e/common/login.cy.js":{"time":5455,"weight":3},"cypress/e2e/lpa/cases/administration.cy.js":{"time":12763,"weight":8},"cypress/e2e/lpa/cases/allocate-case.cy.js":{"time":10187,"weight":7},"cypress/e2e/lpa/cases/change-status.cy.js":{"time":11175,"weight":7},"cypress/e2e/lpa/cases/complaints.cy.js":{"time":18838,"weight":12},"cypress/e2e/lpa/cases/create-event.cy.js":{"time":9663,"weight":6},"cypress/e2e/lpa/cases/create-relationship.cy.js":{"time":11565,"weight":7},"cypress/e2e/lpa/cases/documents.cy.js":{"time":10255,"weight":7},"cypress/e2e/lpa/cases/epa.cy.js":{"time":16680,"weight":11},"cypress/e2e/lpa/cases/investigations.cy.js":{"time":33090,"weight":22},"cypress/e2e/lpa/cases/lpa.cy.js":{"time":20214,"weight":13},"cypress/e2e/lpa/cases/search.cy.js":{"time":8711,"weight":5},"cypress/e2e/lpa/cases/tasks.cy.js":{"time":27771,"weight":19},"cypress/e2e/lpa/cases/timeline.cy.js":{"time":6545,"weight":4},"cypress/e2e/lpa/cases/warnings.cy.js":{"time":16276,"weight":11},"cypress/e2e/lpa/donors/donor.cy.js":{"time":16900,"weight":11},"cypress/e2e/lpa/donors/link.cy.js":{"time":10519,"weight":7},"cypress/e2e/supervision-deputy-hub/navigation.cy.js":{"time":7890,"weight":5},"cypress/e2e/supervision-firm-deputy-hub/navigation.cy.js":{"time":7320,"weight":5},"cypress/e2e/supervision/clients/add-client-search.cy.js":{"time":4590,"weight":3},"cypress/e2e/supervision/clients/add-event.cy.js":{"time":5335,"weight":3},"cypress/e2e/supervision/clients/add-new-client.cy.js":{"time":6830,"weight":4},"cypress/e2e/supervision/clients/allocate-client.cy.js":{"time":9188,"weight":6},"cypress/e2e/supervision/clients/documents/open-document.cy.js":{"time":22622,"weight":15},"cypress/e2e/supervision/clients/edit-client.cy.js":{"time":8723,"weight":6},"cypress/e2e/supervision/clients/edit-finance-details.cy.js":{"time":12551,"weight":8},"cypress/e2e/supervision/clients/record-client-death-confirmation.cy.js":{"time":16050,"weight":11},"cypress/e2e/supervision/clients/risk-score.cy.js":{"time":7858,"weight":5},"cypress/e2e/supervision/clients/visits/add-visit.cy.js":{"time":11686,"weight":8},"cypress/e2e/supervision/clients/visits/edit-visit.cy.js":{"time":9077,"weight":6},"cypress/e2e/supervision/clients/visits/extend-visit.cy.js":{"time":11663,"weight":8},"cypress/e2e/supervision/dashboard/client-dashboard.cy.js":{"time":4083,"weight":2},"cypress/e2e/supervision/dashboard/dashboard.cy.js":{"time":3777,"weight":2},"cypress/e2e/supervision/dashboard/lay-deputy-hub.cy.js":{"time":5895,"weight":4},"cypress/e2e/supervision/dashboard/reassign-client.cy.js":{"time":9535,"weight":6},"cypress/e2e/supervision/deputies/add-deputy.cy.js":{"time":138662,"weight":95},"cypress/e2e/supervision/documents/download-multiple.cy.js":{"time":9525,"weight":6},"cypress/e2e/supervision/help-and-guidance/help-and-guidance.cy.js":{"time":3892,"weight":2},"cypress/e2e/supervision/letters/create-letter.cy.js":{"time":10305,"weight":7},"cypress/e2e/supervision/navigation.cy.js":{"time":4725,"weight":3},"cypress/e2e/supervision/orders/create-order.cy.js":{"time":55169,"weight":37},"cypress/e2e/supervision/search/full-search.cy.js":{"time":6677,"weight":4},"cypress/e2e/supervision/search/search-bar.cy.js":{"time":12720,"weight":8},"cypress/e2e/user-management/my-details.cy.js":{"time":2711,"weight":1},"cypress/e2e/user-management/navigation.cy.js":{"time":5540,"weight":3},"cypress/e2e/workflow/navigation.cy.js":{"time":7450,"weight":5}}
