+++ cypress/support/e2e.js
@@ -4,7 +4,7 @@ require("./deputy-commands");
 require("./lpa-commands");
 require("./visit-commands.js");
 require("./document-commands");
-require("./supervision-commands");
+require("./order-commands");
 
 require("cypress-failed-log");
 require("cy-verify-downloads").addCustomCommand();
